--- a/gemfiles/with_oj.Gemfile
+++ b/gemfiles/with_oj.Gemfile
@@ -1,0 +1,6 @@
+source 'https://rubygems.org'
+
+# Specify your gem's dependencies in lorekeeper.gemspec
+gemspec path: '..'
+
+gem 'oj'