--- a/gemfiles/without_oj.Gemfile
+++ b/gemfiles/without_oj.Gemfile
@@ -1,0 +1,4 @@
+source 'https://rubygems.org'
+
+# Specify your gem's dependencies in lorekeeper.gemspec
+gemspec path: '..'