As suggested in #17 

`Oj` can be made optional. This will only use it if its already required.

For those applications using `Oj` behavior is expected to be 1:1.
For others who choose not to include `Oj`, it will fall back to the default `JSON` from ruby

I also added ruby 2.7 and 3.0 to the build matrix