@jcarres-mdsol 2021-03-02
can you run `rake benchmark` on this branch and on the main branch?  The original usecase for Oj was that it was faster, also calling `defined?` on every log call impacts even applications using Oj