@mjobin-mdsol 2021-03-04
This branch - without Oj

```
$ export BUNDLER_GEMFILE=gemfiles/without_oj.Gemfile
$ rake benchmark
Warming up --------------------------------------
  JSON short content     4.450k i/100ms
Logger short content     7.504k i/100ms
   JSON long content     1.560k i/100ms
 Logger long content     3.743k i/100ms
Calculating -------------------------------------
  JSON short content     44.994k (± 8.1%) i/s -    226.950k in   5.079153s
Logger short content     70.735k (± 7.7%) i/s -    352.688k in   5.017194s
   JSON long content     15.175k (±10.4%) i/s -     76.440k in   5.091825s
 Logger long content     34.234k (±12.5%) i/s -    168.435k in   5.002942s

Comparison:
Logger short content:    70735.4 i/s
  JSON short content:    44994.3 i/s - 1.57x  (± 0.00) slower
 Logger long content:    34233.8 i/s - 2.07x  (± 0.00) slower
   JSON long content:    15174.9 i/s - 4.66x  (± 0.00) slower

i/s means the number of log messages written into a file per second
```

This branch - with Oj
```
$ export BUNDLER_GEMFILE=gemfiles/with_oj.Gemfile
$ rake benchmark
Warming up --------------------------------------
  JSON short content     4.518k i/100ms
Logger short content     7.699k i/100ms
   JSON long content     1.581k i/100ms
 Logger long content     4.659k i/100ms
Calculating -------------------------------------
  JSON short content     45.791k (± 7.2%) i/s -    230.418k in   5.059101s
Logger short content     71.971k (± 9.1%) i/s -    361.853k in   5.073015s
   JSON long content     17.526k (± 6.1%) i/s -     88.536k in   5.070470s
 Logger long content     44.814k (±14.8%) i/s -    223.632k in   5.108709s

Comparison:
Logger short content:    71970.8 i/s
  JSON short content:    45791.3 i/s - 1.57x  (± 0.00) slower
 Logger long content:    44814.3 i/s - 1.61x  (± 0.00) slower
   JSON long content:    17526.3 i/s - 4.11x  (± 0.00) slower

i/s means the number of log messages written into a file per second
```

Master - with Oj

```
$ unset BUNDLER_GEMFILE
$ rake benchmark
Warming up --------------------------------------
  JSON short content     6.461k i/100ms
Logger short content     7.816k i/100ms
   JSON long content     2.523k i/100ms
 Logger long content     4.888k i/100ms
Calculating -------------------------------------
  JSON short content     62.448k (± 7.7%) i/s -    316.589k in   5.101208s
Logger short content     68.237k (± 9.3%) i/s -    343.904k in   5.084572s
   JSON long content     27.806k (± 9.1%) i/s -    138.765k in   5.038996s
 Logger long content     45.572k (±15.4%) i/s -    224.848k in   5.074254s

Comparison:
Logger short content:    68237.2 i/s
  JSON short content:    62448.5 i/s - same-ish: difference falls within error
 Logger long content:    45572.0 i/s - 1.50x  (± 0.00) slower
   JSON long content:    27806.5 i/s - 2.45x  (± 0.00) slower

i/s means the number of log messages written into a file per second
```

Back to `oj-optional` branch, but after removing the `defined?` conditional

```
$ rake benchmark
Warming up --------------------------------------
  JSON short content     4.023k i/100ms
Logger short content     7.750k i/100ms
   JSON long content     1.556k i/100ms
 Logger long content     4.591k i/100ms
Calculating -------------------------------------
  JSON short content     46.283k (± 6.6%) i/s -    233.334k in   5.063980s
Logger short content     74.009k (± 6.6%) i/s -    372.000k in   5.049446s
   JSON long content     17.571k (± 6.9%) i/s -     88.692k in   5.072745s
 Logger long content     49.739k (± 8.6%) i/s -    247.914k in   5.026814s

Comparison:
Logger short content:    74009.1 i/s
 Logger long content:    49739.3 i/s - 1.49x  (± 0.00) slower
  JSON short content:    46283.5 i/s - 1.60x  (± 0.00) slower
   JSON long content:    17571.0 i/s - 4.21x  (± 0.00) slower

i/s means the number of log messages written into a file per second
```