@JordiPolo 2021-03-04
Good thought though, it is good stuff that we try these kind of improvements